Moving functions to organized files based on use
